Mark deprecated styles as unavailable
MOB-2540
