+++ GliaWidgets/Sources/Theme/Chat/Theme.SystemMessageStyle.swift
@@ -40,23 +40,3 @@ public extension Theme {
         }
     }
 }
-
-extension Theme.SystemMessageStyle {
-    func toOldSystemMessageStyle() -> SystemMessageStyle {
-        .init(
-            text: .init(
-                textFont: text.font,
-                textColor: UIColor(hex: text.color),
-                textStyle: text.textStyle,
-                backgroundColor: background.background?.color ?? .clear,
-                cornerRadius: background.cornerRadius,
-                accessibility: .init(
-                    value: accessibility.value,
-                    isFontScalingEnabled: accessibility.isFontScalingEnabled
-                )
-            ),
-            imageFile: imageFile,
-            fileDownload: fileDownload
-        )
-    }
-}
