--- GliaWidgets/Sources/View/Chat/Message/VisitorChatMessageStyle.Accessibility.swift
@@ -1,16 +1,0 @@
-extension VisitorChatMessageStyle {
-    /// Accessibility properties for VisitorChatMessageStyle.
-    public struct Accessibility: Equatable {
-        /// Flag that provides font dynamic type by setting `adjustsFontForContentSizeCategory` for component that supports it.
-        public var isFontScalingEnabled: Bool
-
-        ///
-        /// - Parameter isFontScalingEnabled: Flag that provides font dynamic type by setting `adjustsFontForContentSizeCategory` for component that supports it.
-        public init(isFontScalingEnabled: Bool) {
-            self.isFontScalingEnabled = isFontScalingEnabled
-        }
-
-        /// Accessibility is not supported intentionally.
-        public static let unsupported = Self(isFontScalingEnabled: false)
-    }
-}
